+++ src/plot/both.rs
@@ -12,15 +12,18 @@ use kde;
 use super::{debug_script, scale_time, wait_on_gnuplot};
 use super::{DARK_BLUE, DARK_RED, DEFAULT_FONT, KDE_POINTS, LINEWIDTH, SIZE};
 
+#[cfg_attr(feature = "cargo-clippy", allow(too_many_arguments))]
 pub(crate) fn regression(
     base_data: Data<f64, f64>,
     base_estimates: &Estimates,
     data: Data<f64, f64>,
     estimates: &Estimates,
     id: &str,
-    output_directory: &str,
+    path: String,
+    size: Option<Size>,
+    thumbnail_mode: bool,
 ) {
-    let path = PathBuf::from(format!("{}/{}/both/regression.svg", output_directory, id));
+    let path = PathBuf::from(path);
 
     let max_iters = base_data.x().max().max(data.x().max());
     let max_elapsed = base_data.y().max().max(data.y().max());
@@ -59,10 +62,13 @@ pub(crate) fn regression(
     } = estimates[&Slope];
 
     let mut figure = Figure::new();
+    if !thumbnail_mode {
+        figure.set(Title(id.to_owned()));
+    }
+
     figure
         .set(Font(DEFAULT_FONT))
-        .set(SIZE)
-        .set(Title(id.to_owned()))
+        .set(size.unwrap_or(SIZE))
         .configure(Axis::BottomX, |a| {
             a.configure(Grid::Major, |g| g.show())
                 .set(Label(x_label))
@@ -128,9 +134,11 @@ pub fn pdfs(
     base_avg_times: &Sample<f64>,
     avg_times: &Sample<f64>,
     id: &str,
-    output_directory: &str,
+    path: String,
+    size: Option<Size>,
+    thumbnail_mode: bool,
 ) {
-    let path = PathBuf::from(format!("{}/{}/both/pdf.svg", output_directory, id));
+    let path = PathBuf::from(path);
 
     let (base_xs, base_ys) = kde::sweep(base_avg_times, KDE_POINTS, None);
     let (xs, ys) = kde::sweep(avg_times, KDE_POINTS, None);
@@ -142,10 +150,12 @@ pub fn pdfs(
     let zeros = iter::repeat(0);
 
     let mut figure = Figure::new();
+    if !thumbnail_mode {
+        figure.set(Title(id.to_owned()));
+    }
     figure
         .set(Font(DEFAULT_FONT))
-        .set(SIZE)
-        .set(Title(id.to_owned()))
+        .set(size.unwrap_or(SIZE))
         .configure(Axis::BottomX, |a| {
             a.set(Label(format!("Average time ({}s)", prefix)))
                 .set(ScaleFactor(x_scale))
@@ -155,6 +165,9 @@ pub fn pdfs(
         })
         .configure(Axis::RightY, |a| a.hide())
         .configure(Key, |k| {
+            if thumbnail_mode {
+                k.hide();
+            }
             k.set(Justification::Left)
                 .set(Order::SampleText)
                 .set(Position::Outside(Vertical::Top, Horizontal::Right))